refactor: detour_tile_cache.NavMeshCreateParams 类型别名 + AddTile 零拷贝
- NavMeshCreateParams 改为 detour 包类型别名，消除 bridgeParams 桥接
- detour.NavMeshCreateParams.TileX/TileY/TileLayer int→int32
- AddTile 改为零拷贝（unsafe.Slice），删除 AddTileZeroCopy
- README 添加 int32 字段说明

+++ detour/navmesh_builder.go
@@ -36,9 +36,9 @@ type NavMeshCreateParams struct {
 
 	// Tile Attributes
 	UserID    uint32
-	TileX     int
-	TileY     int
-	TileLayer int
+	TileX     int32
+	TileY     int32
+	TileLayer int32
 	Bmin      [3]float32
 	Bmax      [3]float32
 
@@ -420,9 +420,9 @@ func CreateNavMeshData(params *NavMeshCreateParams) ([]byte, int, bool) {
 	header := &MeshHeader{
 		Magic:           NavMeshMagic,
 		Version:         NavMeshVersion,
-		X:               int32(params.TileX),
-		Y:               int32(params.TileY),
-		Layer:           int32(params.TileLayer),
+		X:               params.TileX,
+		Y:               params.TileY,
+		Layer:           params.TileLayer,
 		UserID:          params.UserID,
 		PolyCount:       int32(totPolyCount),
 		VertCount:       int32(totVertCount),
